Distance.jl

Types and Enums

Distances.DistanceType — Type

This enum is used to select the distance calculation type. We converted the name into a string, and the magic happens! :D

DistanceType Possibilities:

Euclidean, Cityblock, TotalVariation, Chebyshev,
Jaccard, BrayCurtis, SpanNormDist

Examples:

Euclidean::DistanceType
Cityblock::DistanceType

Functions

Distances.GetDistance — Method

GetDistance(distanceType::DistanceType,
v1::AbstractArray{<: Number}, v2::AbstractArray{<: Number})

This function compute distance between two numeric vectors. You can use diferents distance calculations selecting with "DistanceType".

Examples:

GetDistance(Euclidean::DistanceType, [2,2], [2,4])
GetDistance(Cityblock::DistanceType, [2,2], [2,4])
GetDistance(TotalVariation::DistanceType, [2,2], [2,4])